Move some more Arbitrary to cardano-protocol
