+++ eras/alonzo/impl/cardano-ledger-alonzo.cabal
@@ -1,6 +1,6 @@
 cabal-version: 3.0
 name: cardano-ledger-alonzo
-version: 1.16.0.0
+version: 1.16.1.0
 license: Apache-2.0
 maintainer: operations@iohk.io
 author: IOHK
