+++ libs/cardano-protocol-tpraos/testlib/Test/Cardano/Protocol/TPraos/Arbitrary.hs
@@ -27,8 +27,8 @@ import Cardano.Ledger.Core
 import Cardano.Protocol.Crypto (Crypto (KES, VRF), StandardCrypto)
 import Cardano.Protocol.TPraos.API (ChainDepState, PraosCrypto)
 import Cardano.Protocol.TPraos.BlockHeader (
-  BHBody (BHBody),
-  BHeader (BHeader),
+  BHBody,
+  BHeader,
  )
 import Cardano.Protocol.TPraos.OCert (KESPeriod (KESPeriod))
 import Cardano.Protocol.TPraos.Rules.Overlay (OBftSlot)
@@ -79,19 +79,6 @@ instance Arbitrary OBftSlot where
   arbitrary = genericArbitraryU
   shrink = genericShrink
 
-instance
-  ( Crypto c
-  , VRF.Signable (VRF c) ~ SignableRepresentation
-  , KES.Signable (KES c) ~ SignableRepresentation
-  ) =>
-  Arbitrary (BHeader c)
-  where
-  arbitrary = do
-    bhBody <- arbitrary
-    hotKey <- arbitrary
-    let sig = KES.unsoundPureSignedKES () 1 bhBody hotKey
-    pure $ BHeader bhBody sig
-
 genBHeader ::
   ( VRF.Signable (VRF c) Seed
   , KES.Signable (KES c) (BHBody c)
@@ -114,41 +101,6 @@ genBHeader aiks = do
       bhBody =
         mkBHBody protVer prevHash allPoolKeys slotNo blockNo epochNonce oCert bodySize bodyHash
   return $ mkBHeader allPoolKeys kesPeriod keyRegKesPeriod bhBody
-
-instance
-  ( Crypto c
-  , VRF.Signable (VRF c) ~ SignableRepresentation
-  ) =>
-  Arbitrary (BHBody c)
-  where
-  arbitrary =
-    BHBody
-      <$> arbitrary
-      <*> arbitrary
-      <*> arbitrary
-      <*> arbitrary
-      <*> arbitrary
-      <*> arbitrary
-      <*> arbitrary
-      <*> arbitrary
-      <*> arbitrary
-      <*> arbitrary
-      <*> arbitrary
-
-instance
-  ( Crypto c
-  , EraBlockBody era
-  , KES.Signable (KES c) ~ SignableRepresentation
-  , VRF.Signable (VRF c) ~ SignableRepresentation
-  , Arbitrary (Tx TopTx era)
-  , Arbitrary (BlockBody era)
-  ) =>
-  Arbitrary (Block (BHeader c) era)
-  where
-  arbitrary =
-    Block
-      <$> arbitrary
-      <*> arbitrary
 
 -- | Use supplied keys to generate a Block.
 genBlock ::
