+++ libs/cardano-protocol/cardano-protocol.cabal
@@ -1,6 +1,6 @@
 cabal-version: 3.0
 name: cardano-protocol
-version: 0.1.0.0
+version: 0.1.1.0
 license: Apache-2.0
 maintainer: operations@iohk.io
 author: IOHK
